refactor: move addPlugin to FxCore.declarativeAgent.ts
