+++ packages/fx-core/src/core/FxCore.ts
@@ -2026,454 +2026,6 @@ export class FxCore extends FxCoreDeclarativeAgentPart {
     return ok(projectTypeRes);
   }
 
-  /**
-   * Add plugin
-   */
-  @hooks([
-    ErrorContextMW({ component: "FxCore", stage: Stage.addPlugin }),
-    ErrorHandlerMW,
-    QuestionMW("addPlugin"),
-    ConcurrentLockerMW,
-  ])
-  async addPlugin(inputs: Inputs): Promise<Result<undefined | any, FxError>> {
-    if (!inputs.projectPath) {
-      throw new Error("projectPath is undefined"); // should never happen
-    }
-
-    const context = createContext();
-    const teamsManifestPath = inputs[QuestionNames.ManifestPath];
-    const appPackageFolder = path.dirname(teamsManifestPath);
-    const isGenerateFromApiSpec =
-      inputs[QuestionNames.ActionType] === ActionStartOptions.apiSpec().id;
-    const isGenerateFromMCP = inputs[QuestionNames.ActionType] === ActionStartOptions.mcp().id;
-
-    // validate the project is valid for adding plugin
-    const manifestRes = await manifestUtils._readAppManifest(teamsManifestPath);
-    if (manifestRes.isErr()) {
-      return err(manifestRes.error);
-    }
-
-    const teamsManifest = manifestRes.value;
-    const declarativeGpt = teamsManifest.copilotExtensions
-      ? teamsManifest.copilotExtensions.declarativeCopilots?.[0]
-      : teamsManifest.copilotAgents?.declarativeAgents?.[0];
-    if (!declarativeGpt?.file) {
-      return err(
-        AppStudioResultFactory.UserError(
-          AppStudioError.TeamsAppRequiredPropertyMissingError.name,
-          AppStudioError.TeamsAppRequiredPropertyMissingError.message(
-            "declarativeAgents",
-            teamsManifestPath
-          )
-        )
-      );
-    }
-    const gptManifestFilePathRes = await copilotGptManifestUtils.getManifestPath(teamsManifestPath);
-    if (gptManifestFilePathRes.isErr()) {
-      return err(gptManifestFilePathRes.error);
-    }
-
-    const declarativeCopilotManifestPath = gptManifestFilePathRes.value;
-
-    const declarativeCopilotManifesRes = await copilotGptManifestUtils.readCopilotGptManifestFile(
-      declarativeCopilotManifestPath
-    );
-    if (declarativeCopilotManifesRes.isErr()) {
-      return err(declarativeCopilotManifesRes.error);
-    }
-
-    const declarativeCopilotManifest = declarativeCopilotManifesRes.value;
-    let confirmMessage = getLocalizedString(
-      "core.addApi.confirm",
-      path.relative(inputs.projectPath, appPackageFolder)
-    );
-
-    // Will be used if generating from API spec
-    let specParser: SpecParser | undefined = undefined;
-    let authNameAndSchemes: { authName: string; authScheme: AuthType }[] = [];
-
-    if (isGenerateFromApiSpec) {
-      specParser = new SpecParser(
-        inputs[QuestionNames.ApiSpecLocation].trim(),
-        getParserOptions(ProjectType.Copilot, true)
-      );
-      const listResult = await listAPIInfo(inputs[QuestionNames.ApiSpecLocation].trim());
-      authNameAndSchemes = this.parseAuthNameAndScheme(listResult, inputs);
-
-      if (authNameAndSchemes.length > 0) {
-        const doesLocalYamlPathExists = await fs.pathExists(
-          path.join(inputs.projectPath, MetadataV3.localConfigFile)
-        );
-        confirmMessage = doesLocalYamlPathExists
-          ? getLocalizedString(
-              "core.addApi.confirm.localTeamsYaml",
-              path.relative(inputs.projectPath, appPackageFolder),
-              MetadataV4.localConfigFile,
-              MetadataV4.configFile
-            )
-          : getLocalizedString(
-              "core.addApi.confirm.teamsYaml",
-              path.relative(inputs.projectPath, appPackageFolder),
-              MetadataV4.configFile
-            );
-      }
-    }
-
-    // confirm
-
-    const confirmRes = await context.userInteraction.showMessage(
-      "warn",
-      confirmMessage,
-      true,
-      getLocalizedString("core.addApi.continue")
-    );
-
-    if (confirmRes.isErr()) {
-      return err(confirmRes.error);
-    } else if (confirmRes.value !== getLocalizedString("core.addApi.continue")) {
-      return err(new UserCancelError());
-    }
-
-    // find the next available action id
-    let actionId = "";
-    let suffix = 1;
-    actionId = `action_${suffix}`;
-    const existingActionIds = declarativeCopilotManifest.actions?.map((action) => action.id);
-    while (existingActionIds?.includes(actionId)) {
-      suffix += 1;
-      actionId = `action_${suffix}`;
-    }
-
-    let destinationPluginManifestPath: string;
-    // generate files
-    if (isGenerateFromApiSpec && specParser) {
-      destinationPluginManifestPath =
-        await copilotGptManifestUtils.getDefaultNextAvailablePluginManifestPath(
-          appPackageFolder,
-          undefined
-        );
-      const destinationApiSpecPath = await pluginManifestUtils.getDefaultNextAvailableApiSpecPath(
-        inputs[QuestionNames.ApiSpecLocation].trim(),
-        path.join(appPackageFolder, DefaultApiSpecFolderName)
-      );
-
-      const generateRes = await generateFromApiSpec(
-        specParser,
-        teamsManifestPath,
-        inputs,
-        context,
-        Stage.addPlugin,
-        ProjectType.Copilot,
-        {
-          destinationApiSpecFilePath: destinationApiSpecPath,
-          pluginManifestFilePath: destinationPluginManifestPath,
-        },
-        inputs[QuestionNames.ApiSpecLocation].trim()
-      );
-      if (generateRes.isErr()) {
-        return err(generateRes.error);
-      }
-
-      const warnings = generateRes.value.warnings;
-      if (warnings && warnings.length > 0) {
-        const warnSummary = await generateScaffoldingSummary(
-          warnings,
-          manifestRes.value,
-          path.relative(inputs.projectPath, destinationApiSpecPath),
-          path.relative(inputs.projectPath, destinationPluginManifestPath),
-          inputs.projectPath
-        );
-        context.logProvider.info(warnSummary + "\n");
-      }
-
-      const addActionRes = await copilotGptManifestUtils.addAction(
-        declarativeCopilotManifestPath,
-        actionId,
-        normalizePath(path.relative(appPackageFolder, destinationPluginManifestPath), true)
-      );
-      if (addActionRes.isErr()) {
-        return err(addActionRes.error);
-      }
-
-      for (const authNameAndScheme of authNameAndSchemes) {
-        await this.updateAuthActionInYaml(
-          authNameAndScheme.authName,
-          authNameAndScheme.authScheme,
-          inputs.projectPath,
-          destinationApiSpecPath,
-          destinationPluginManifestPath
-        );
-      }
-    } else if (isGenerateFromMCP) {
-      // MCP action: create ai-plugin.json with MCP runtime config
-      const mcpServerUrl = inputs[QuestionNames.MCPForDAServerUrl];
-      const toolsFilePath = inputs[QuestionNames.MCPToolsFilePath];
-      const mcpWarnings: Warning[] = [];
-
-      // Validate: --mcp-da-server-url is required for MCP actions
-      if (!mcpServerUrl) {
-        return err(
-          new UserError(
-            Stage.addPlugin,
-            "MissingMCPServerUrl",
-            getLocalizedString("core.MCPForDA.missingServerUrl"),
-            getLocalizedString("core.MCPForDA.missingServerUrl")
-          )
-        );
-      }
-
-      // Load tools from file if provided and not yet loaded
-      const existingTools = inputs[QuestionNames.MCPForDAAvailableTools];
-      if (toolsFilePath && (!existingTools || existingTools.length === 0)) {
-        try {
-          const { readMCPToolsFromFile } = await import("../component/utils/mcpToolFetcher");
-          const fileTools = await readMCPToolsFromFile(toolsFilePath);
-          inputs[QuestionNames.MCPForDAAvailableTools] = fileTools;
-          if (!inputs[QuestionNames.MCPForDAPreFetchTools]) {
-            inputs[QuestionNames.MCPForDAPreFetchTools] = fileTools.map((t: any) => t.name);
-          }
-        } catch {
-          mcpWarnings.push({
-            type: "mcpToolsFileReadError",
-            content: getLocalizedString("core.MCPForDA.toolsFileReadError", toolsFilePath),
-          });
-        }
-      }
-
-      // Probe auth if tools were loaded from file but auth not yet detected
-      if (
-        inputs[QuestionNames.MCPForDAAvailableTools]?.length > 0 &&
-        !inputs[QuestionNames.MCPForDAAuth] &&
-        mcpServerUrl
-      ) {
-        try {
-          const { probeMCPServerAuth } = await import("../component/utils/mcpToolFetcher");
-          const authProbe = await probeMCPServerAuth(mcpServerUrl);
-          if (authProbe.requiresAuth) {
-            inputs[QuestionNames.MCPForDAAuth] = "OAuthPluginVault";
-            if (authProbe.authMetadataUrl) {
-              inputs[QuestionNames.MCPForDAAuthMetadataUrl] = authProbe.authMetadataUrl;
-            }
-          }
-        } catch {
-          // Auth probe failed — continue without auth
-        }
-      }
-
-      // Auto-fetch tools if not yet loaded (CLI flow for no-auth servers)
-      const currentTools = inputs[QuestionNames.MCPForDAAvailableTools];
-      if ((!currentTools || currentTools.length === 0) && mcpServerUrl) {
-        try {
-          const { fetchMCPTools } = await import("../component/utils/mcpToolFetcher");
-          const result = await fetchMCPTools(mcpServerUrl);
-          if (!result.requiresAuth && result.tools.length > 0) {
-            inputs[QuestionNames.MCPForDAAvailableTools] = result.tools;
-            if (!inputs[QuestionNames.MCPForDAPreFetchTools]) {
-              inputs[QuestionNames.MCPForDAPreFetchTools] = result.tools.map((t) => t.name);
-            }
-          } else if (result.requiresAuth) {
-            // Store auth metadata for later use
-            inputs[QuestionNames.MCPForDAAuth] = "OAuthPluginVault";
-            if (result.authMetadataUrl) {
-              inputs[QuestionNames.MCPForDAAuthMetadataUrl] = result.authMetadataUrl;
-            }
-            mcpWarnings.push({
-              type: "mcpAuthRequired",
-              content: getLocalizedString("core.MCPForDA.authRequired", mcpServerUrl),
-            });
-          } else {
-            mcpWarnings.push({
-              type: "mcpNoToolsFetched",
-              content: getLocalizedString("core.MCPForDA.noToolsFetched", mcpServerUrl),
-            });
-          }
-        } catch {
-          mcpWarnings.push({
-            type: "mcpFetchError",
-            content: getLocalizedString("core.MCPForDA.fetchError", mcpServerUrl),
-          });
-        }
-      }
-
-      // If no tools available (auth required or fetch failed), skip creating plugin — just warn
-      const mcpTools = inputs[QuestionNames.MCPForDAAvailableTools];
-      const mcpToolsSelected = inputs[QuestionNames.MCPForDAPreFetchTools] || [];
-      if (!mcpTools || mcpTools.length === 0 || mcpToolsSelected.length === 0) {
-        if (mcpWarnings.length > 0) {
-          for (const warning of mcpWarnings) {
-            context.logProvider.warning(warning.content);
-          }
-        }
-        return ok(undefined);
-      }
-
-      destinationPluginManifestPath =
-        await copilotGptManifestUtils.getDefaultNextAvailablePluginManifestPath(
-          appPackageFolder,
-          undefined
-        );
-
-      const pluginManifest: any = {
-        $schema: "https://developer.microsoft.com/json-schemas/copilot/plugin/v2.4/schema.json",
-        schema_version: "v2.4",
-        name_for_human: "MCP Action",
-        description_for_human: "Action powered by MCP server",
-        contact_email: "publisher-email@example.com",
-        namespace: "mcpAction",
-        functions: [],
-        runtimes: [],
-      };
-
-      if (mcpTools && mcpTools.length > 0 && mcpToolsSelected.length > 0) {
-        const selectedTools = mcpTools.filter((tool: any) => mcpToolsSelected.includes(tool.name));
-
-        pluginManifest.functions = selectedTools.map((tool: any) => ({
-          name: tool.name,
-          description: tool.description || "",
-        }));
-
-        // Write mcp-tools file with full tool definitions (no field filtering)
-        const mcpToolsFile = `mcp-tools-${suffix}.json`;
-        const mcpToolsOutputPath = path.join(appPackageFolder, mcpToolsFile);
-        await fs.writeJSON(mcpToolsOutputPath, { tools: selectedTools }, { spaces: 4 });
-
-        pluginManifest.runtimes = [
-          {
-            type: "RemoteMCPServer",
-            spec: {
-              url: mcpServerUrl,
-              mcp_tool_description: {
-                file: mcpToolsFile,
-              },
-            },
-            run_for_functions: pluginManifest.functions.map((f: any) => f.name),
-          },
-        ];
-
-        // Auth handling
-        const mcpAuth = inputs[QuestionNames.MCPForDAAuth];
-        if (mcpAuth === "OAuthPluginVault") {
-          const authType = inputs[QuestionNames.MCPForDAAuthType];
-          if (!authType) {
-            return err(
-              new UserError(
-                Stage.addPlugin,
-                "MissingMCPAuthType",
-                getLocalizedString("core.MCPForDA.missingAuthType"),
-                getLocalizedString("core.MCPForDA.missingAuthType")
-              )
-            );
-          }
-
-          const registrationId = `MCP_DA_AUTH_ID_${actionId.toUpperCase()}`;
-          pluginManifest.runtimes[0].auth = {
-            type: "OAuthPluginVault",
-            reference_id: `$\{\{${registrationId}\}\}`,
-          };
-
-          // Resolve OAuth metadata and inject oauth/register into m365agents.yml
-          try {
-            let authorizationUrl: string | undefined;
-            let tokenUrl: string | undefined;
-            let refreshUrl: string | undefined;
-
-            if (authType === "oauth") {
-              const { resolveMCPOAuthMetadata } = await import("../component/utils/mcpToolFetcher");
-              const metadata = await resolveMCPOAuthMetadata(
-                inputs[QuestionNames.MCPForDAAuthMetadataUrl]
-              );
-              authorizationUrl = metadata.authorizationUrl;
-              tokenUrl = metadata.tokenUrl;
-              refreshUrl = metadata.refreshUrl;
-            }
-
-            const { ActionInjector } = await import("../component/configManager/actionInjector");
-            const ymlPath = pathUtils.getYmlFilePath(inputs.projectPath);
-            if (ymlPath) {
-              await ActionInjector.injectCreateOAuthActionForMCP(
-                ymlPath,
-                authType,
-                actionId,
-                registrationId,
-                mcpServerUrl,
-                authorizationUrl,
-                tokenUrl,
-                refreshUrl
-              );
-            }
-          } catch (error: any) {
-            mcpWarnings.push({
-              type: "mcpAuthMetadataError",
-              content: getLocalizedString(
-                "core.MCPForDA.mcpAuthMetadataMissingError",
-                error.message
-              ),
-            });
-          }
-        }
-      }
-
-      await fs.ensureFile(destinationPluginManifestPath);
-      await fs.writeJSON(destinationPluginManifestPath, pluginManifest, { spaces: 4 });
-
-      const addActionRes = await copilotGptManifestUtils.addAction(
-        declarativeCopilotManifestPath,
-        actionId,
-        normalizePath(path.relative(appPackageFolder, destinationPluginManifestPath), true)
-      );
-      if (addActionRes.isErr()) {
-        return err(addActionRes.error);
-      }
-
-      // Surface warnings for CLI
-      if (mcpWarnings.length > 0) {
-        for (const warning of mcpWarnings) {
-          context.logProvider.warning(warning.content);
-        }
-      }
-    } else {
-      const addPluginRes = await addExistingPlugin(
-        declarativeCopilotManifestPath,
-        inputs[QuestionNames.PluginManifestFilePath].trim(),
-        inputs[QuestionNames.PluginOpenApiSpecFilePath].trim(),
-        actionId,
-        context,
-        Stage.addPlugin
-      );
-
-      if (addPluginRes.isErr()) {
-        return err(addPluginRes.error);
-      }
-      destinationPluginManifestPath = addPluginRes.value.destinationPluginManifestPath;
-      const warningMessage = outputScaffoldingWarningMessage(addPluginRes.value.warnings);
-      context.logProvider.info(warningMessage);
-    }
-
-    if (inputs.platform === Platform.VSCode) {
-      const successMessage = getLocalizedString("core.addPlugin.success.vsc", actionId);
-      const viewPluginManifest = getLocalizedString("core.addPlugin.success.viewPluginManifest");
-      void context.userInteraction
-        .showMessage("info", successMessage, false, viewPluginManifest)
-        .then((userRes) => {
-          if (userRes.isOk() && userRes.value === viewPluginManifest) {
-            context.telemetryReporter.sendTelemetryEvent(
-              TelemetryEvent.ViewPluginManifestAfterAdded
-            );
-            void TOOLS?.ui?.openFile?.(destinationPluginManifestPath);
-          }
-        });
-    } else {
-      const successMessage = getLocalizedString(
-        "core.addPlugin.success",
-        actionId,
-        destinationPluginManifestPath
-      );
-      void context.userInteraction.showMessage("info", successMessage, false);
-    }
-
-    return ok(undefined);
-  }
-
   @hooks([
     ErrorContextMW({ component: "FxCore", stage: Stage.RegeneratePlugin }),
     ErrorHandlerMW,
@@ -3273,7 +2825,7 @@ export class FxCore extends FxCoreDeclarativeAgentPart {
     return await generateConfigFiles(inputs);
   }
 
-  private async updateAuthActionInYaml(
+  protected async updateAuthActionInYaml(
     authName: string | undefined,
     authScheme: AuthType | undefined,
     projectPath: string,
@@ -3304,7 +2856,7 @@ export class FxCore extends FxCoreDeclarativeAgentPart {
     }
   }
 
-  private parseAuthNameAndScheme(
+  protected parseAuthNameAndScheme(
     listResult: ListAPIResult,
     inputs: Inputs
   ): { authName: string; authScheme: AuthType }[] {
